[ADD] -- logout and interceptor

+++ src/pages/accounts/Logout.js
@@ -1,27 +1,27 @@
 import axios from "axios";
 import React, { useEffect } from "react";
 import { baseURL } from "../../services/user.services";
+import { axiosapiSecure } from "../../interceptor/axios";
 
 function Logout() {
   useEffect(() => {
-    (async () => {
-      try {
-        debugger;
-        const data = await axios.post(
+    if (localStorage.getItem("access_token") === null) {
+      window.location.href = "/login";
+    } else {
+      axiosapiSecure
+        .post(
           `${baseURL}logout/`,
           {
             refresh_token: localStorage.getItem("refresh_token"),
           },
-          { headers: { "Content-Type": "application/json" } },
-          { withCredentials: true }
-        );
-        localStorage.clear();
-        axios.defaults.headers.common["Authorization"] = null;
-        window.location.href = "/login";
-      } catch (e) {
-        console.log("logout not working", e);
-      }
-    })();
+        )
+        .then((response) => {
+          localStorage.clear();
+          axios.defaults.headers.common["Authorization"] = null;
+          window.location.href = "/login";
+        })
+        .catch((response) => console.log(response));
+    }
   }, []);
   return <div></div>;
 }
